refactor(app): decompose ManageLLM into overview panel and composables (#397)
Split ManageLLM.vue (938 -> 273 lines) by extracting the overview tab content
into LlmOverviewPanel.vue and the regeneration-job orchestration
(useLlmRegenerationJobs.ts), cache-status presentation (useLlmCacheOverview.ts),
and hash tab navigation (useLlmAdminTabs.ts). Session-scoped job persistence,
resume-on-return behavior, child-job polling, and all status/feedback copy are
preserved verbatim.

Advances #397. Part of #346.
